@@ -1,11 +1,10 @@
 <?php
 
-namespace Intervention\Gif;
+namespace Intervention\Gif\Blocks;
 
-use Intervention\Gif\Contracts\SpecialPurposeBlock;
-use Intervention\Gif\DataSubBlock;
+use Intervention\Gif\AbstractExtension;
 
-class ApplicationExtension extends AbstractExtension implements SpecialPurposeBlock
+class ApplicationExtension extends AbstractExtension
 {
     public const LABEL = "\xFF";
 
