@@ -2,8 +2,8 @@
 
 namespace Intervention\Gif\Decoder;
 
-use Intervention\Gif\Color;
-use Intervention\Gif\ColorTable;
+use Intervention\Gif\Blocks\Color;
+use Intervention\Gif\Blocks\ColorTable;
 
 class ColorTableDecoder extends AbstractDecoder
 {
