@@ -2,7 +2,7 @@
 
 namespace Intervention\Gif\Encoder;
 
-use Intervention\Gif\TableBasedImage as TableBasedImage;
+use Intervention\Gif\Blocks\TableBasedImage;
 
 class TableBasedImageEncoder extends AbstractEncoder
 {
@@ -16,29 +16,12 @@ class TableBasedImageEncoder extends AbstractEncoder
         $this->source = $source;
     }
 
-    /**
-     * Encode current source
-     *
-     * @return string
-     */
     public function encode(): string
     {
         return implode('', [
-            $this->source->getDescriptor()->encode(),
-            $this->encodeLocalColorTable(),
-            $this->source->getData()->encode()
+            $this->source->getImageDescriptor()->encode(),
+            $this->source->getColorTable() ? $this->source->getColorTable()->encode() : '',
+            $this->source->getImageData()->encode(),
         ]);
-    }
-
-    /**
-     * Encode local color table if available
-     *
-     * @return string
-     */
-    protected function encodeLocalColorTable(): string
-    {
-        $table = $this->source->getColorTable();
-
-        return $table ? $table->encode() : '';
     }
 }
