@@ -19,26 +19,29 @@ class BuilderTest extends BaseTestCase
         $this->assertMatchesRegularExpression('/^\x47\x49\x46\x38(\x37|\x39)\x61/', $builder->encode());
     }
 
-    public function testCanvasDefaultLoops()
+    public function testCanvas()
     {
         $builder = Builder::canvas(320, 240);
         $this->assertInstanceOf(Builder::class, $builder);
         $gif = $builder->getGifDataStream();
-        $this->assertEquals(320, $gif->getLogicalScreen()->getDescriptor()->getWidth());
-        $this->assertEquals(240, $gif->getLogicalScreen()->getDescriptor()->getHeight());
-        $this->assertEquals(0, $gif->getMainApplicationExtension()->getLoops());
+        $this->assertEquals(320, $gif->getLogicalScreenDescriptor()->getWidth());
+        $this->assertEquals(240, $gif->getLogicalScreenDescriptor()->getHeight());
     }
 
     public function testCanvasOneLoops()
     {
-        $builder = Builder::canvas(320, 240, 1);
+        $builder = Builder::canvas(320, 240);
+        $builder->addFrame(__DIR__ . '/images/red.gif', 0.25, 1, 2);
+        $builder->setLoops(1);
         $gif = $builder->getGifDataStream();
         $this->assertNull($gif->getMainApplicationExtension());
     }
 
     public function testCanvasMultipleLoops()
     {
-        $builder = Builder::canvas(320, 240, 10);
+        $builder = Builder::canvas(320, 240);
+        $builder->addFrame(__DIR__ . '/images/red.gif', 0.25, 1, 2);
+        $builder->setLoops(10);
         $gif = $builder->getGifDataStream();
         $this->assertEquals(10, $gif->getMainApplicationExtension()->getLoops());
     }
@@ -48,9 +51,9 @@ class BuilderTest extends BaseTestCase
         $builder = Builder::canvas(320, 240);
         $result = $builder->addFrame(__DIR__ . '/images/red.gif', 0.25, 1, 2);
         $this->assertInstanceOf(Builder::class, $result);
-        $blocks = $result->getGifDataStream()->getGraphicBlocks();
-        $this->assertEquals(25, $blocks[0]->getGraphicControlExtension()->getDelay());
-        $this->assertEquals(1, $blocks[0]->getGraphicRenderingBlock()->getDescriptor()->getLeft());
-        $this->assertEquals(2, $blocks[0]->getGraphicRenderingBlock()->getDescriptor()->getTop());
+        $gif = $builder->getGifDataStream();
+        $this->assertEquals(25, $gif->getFirstFrame()->getGraphicControlExtension()->getDelay());
+        $this->assertEquals(1, $gif->getFirstFrame()->getImageDescriptor()->getLeft());
+        $this->assertEquals(2, $gif->getFirstFrame()->getImageDescriptor()->getTop());
     }
 }
