@@ -2,10 +2,7 @@
 
 namespace Intervention\Gif\Test;
 
-use Intervention\Gif\Color;
-use Intervention\Gif\ColorTable;
-use Intervention\Gif\Exception\DecodingException;
-use Intervention\Gif\Header;
+use Intervention\Gif\Blocks\Header;
 
 class HeaderTest extends BaseTestCase
 {
